+++ __tests__/dagre/layer-test.ts
@@ -1,0 +1,741 @@
+import { Layouts } from '../../src'
+import G6 from '@antv/g6';
+
+
+const data ={
+  "nodes": [
+      {
+          "id": "node_0_0_ss",
+          "layer": 0.5,
+          "label": "node0",
+          "custom": {}
+      },
+      {
+          "id": "node_1_0_ss",
+          "layer": 1.5,
+          "label": "node1",
+          "custom": {}
+      },
+      {
+          "id": "node_1_1_ss",
+          "layer": 1.5,
+          "label": "node2",
+          "custom": {}
+      },
+      {
+          "id": "node_1_2_ss",
+          "layer": 1.5,
+          "label": "node3",
+          "custom": {}
+      },
+      {
+          "id": "node_1_3_ss",
+          "layer": 1.5,
+          "label": "node4",
+          "custom": {}
+      },
+      {
+          "id": "node_2_0_ss",
+          "layer": 2,
+          "label": "node5",
+          "custom": {}
+      },
+      {
+          "id": "node_2_1_ss",
+          "layer": 2,
+          "label": "node6",
+          "custom": {}
+      },
+      {
+          "id": "node_2_2_ss",
+          "layer": 2,
+          "label": "node7",
+          "custom": {}
+      },
+      {
+          "id": "node_2_3_ss",
+          "layer": 2,
+          "label": "node8",
+          "custom": {}
+      },
+      {
+          "id": "node_3_0_ss",
+          "layer": 1,
+          "label": "node9",
+          "custom": {}
+      },
+      {
+          "id": "node_3_1_ss",
+          "layer": 1,
+          "label": "node10",
+          "custom": {}
+      },
+      {
+          "id": "node_3_2_ss",
+          "layer": 1,
+          "label": "node11",
+          "custom": {}
+      },
+      {
+          "id": "node_3_3_ss",
+          "layer": 1,
+          "label": "node12",
+          "custom": {}
+      },
+      {
+          "id": "node_0_0_cjdx",
+          "layer": 0.5,
+          "label": "node13",
+          "custom": {}
+      },
+      {
+          "id": "node_0_1_cjdx",
+          "layer": 0.5,
+          "label": "node14",
+          "custom": {}
+      },
+      {
+          "id": "node_0_2_cjdx",
+          "layer": 0.5,
+          "label": "node15",
+          "custom": {}
+      },
+      {
+          "id": "node_0_3_cjdx",
+          "layer": 0.5,
+          "label": "node16",
+          "custom": {}
+      },
+      {
+          "id": "node_0_4_cjdx",
+          "layer": 0.5,
+          "label": "node17",
+          "custom": {}
+      },
+      {
+          "id": "node_0_5_cjdx",
+          "layer": 0.5,
+          "label": "node18",
+          "custom": {}
+      },
+      {
+          "id": "node_0_6_cjdx",
+          "layer": 0.5,
+          "label": "node19",
+          "custom": {}
+      },
+      {
+          "id": "node_0_7_cjdx",
+          "layer": 0.5,
+          "label": "node20",
+          "custom": {}
+      },
+      {
+          "id": "node_0_8_cjdx",
+          "layer": 0.5,
+          "label": "node21",
+          "custom": {}
+      },
+      {
+          "id": "node_1_0_cjdx",
+          "layer": 2,
+          "label": "node22",
+          "custom": {}
+      },
+      {
+          "id": "node_1_1_cjdx",
+          "layer": 2,
+          "label": "node23",
+          "custom": {}
+      },
+      {
+          "id": "node_1_2_cjdx",
+          "layer": 2,
+          "label": "node24",
+          "custom": {}
+      },
+      {
+          "id": "node_1_3_cjdx",
+          "layer": 2,
+          "label": "node25",
+          "custom": {}
+      },
+      {
+          "id": "node_1_4_cjdx",
+          "layer": 2,
+          "label": "node26",
+          "custom": {}
+      },
+      {
+          "id": "node_1_5_cjdx",
+          "layer": 2,
+          "label": "node27",
+          "custom": {}
+      },
+      {
+          "id": "node_1_6_cjdx",
+          "layer": 2,
+          "label": "node28",
+          "custom": {}
+      },
+      {
+          "id": "node_1_7_cjdx",
+          "layer": 2,
+          "label": "node29",
+          "custom": {}
+      },
+      {
+          "id": "node_1_8_cjdx",
+          "layer": 2,
+          "label": "node30",
+          "custom": {}
+      },
+      {
+          "id": "node_2_0_cjdx",
+          "layer": 1,
+          "label": "node31",
+          "custom": {}
+      },
+      {
+          "id": "node_2_1_cjdx",
+          "layer": 1,
+          "label": "node32",
+          "custom": {}
+      },
+      {
+          "id": "node_2_2_cjdx",
+          "layer": 1,
+          "label": "node33",
+          "custom": {}
+      },
+      {
+          "id": "node_2_3_cjdx",
+          "layer": 1,
+          "label": "node34",
+          "custom": {}
+      },
+      {
+          "id": "node_2_4_cjdx",
+          "layer": 1,
+          "label": "node35",
+          "custom": {}
+      },
+      {
+          "id": "node_2_5_cjdx",
+          "layer": 1,
+          "label": "node36",
+          "custom": {}
+      },
+      {
+          "id": "node_2_6_cjdx",
+          "layer": 1,
+          "label": "node37",
+          "custom": {}
+      },
+      {
+          "id": "node_2_7_cjdx",
+          "layer": 1,
+          "label": "node38",
+          "custom": {}
+      },
+      {
+          "id": "node_2_8_cjdx",
+          "layer": 1,
+          "label": "node39",
+          "custom": {}
+      },
+      {
+          "id": "node_1_4_ss",
+          "layer": 1.5,
+          "label": "node40",
+          "custom": {}
+      },
+      {
+          "id": "node_2_4_ss",
+          "layer": 2,
+          "label": "node41",
+          "custom": {}
+      },
+      {
+          "id": "node_3_4_ss",
+          "layer": 1,
+          "label": "node42",
+          "custom": {}
+      },
+      {
+          "id": "node_0_0_gy",
+          "layer": 0.5,
+          "label": "node43",
+          "custom": {}
+      },
+      {
+          "id": "node_0_1_gy",
+          "layer": 0.5,
+          "label": "node44",
+          "custom": {}
+      },
+      {
+          "id": "node_0_2_gy",
+          "layer": 0.5,
+          "label": "node45",
+          "custom": {}
+      },
+      {
+          "id": "node_1_0_gy",
+          "layer": 1.5,
+          "label": "node46",
+          "custom": {}
+      },
+      {
+          "id": "node_1_1_gy",
+          "layer": 1.5,
+          "label": "node47",
+          "custom": {}
+      },
+      {
+          "id": "node_1_2_gy",
+          "layer": 1.5,
+          "label": "node48",
+          "custom": {}
+      },
+      {
+          "id": "node_1_3_gy",
+          "layer": 1.5,
+          "label": "node49",
+          "custom": {}
+      },
+      {
+          "id": "node_2_0_gy",
+          "layer": 2,
+          "label": "node50",
+          "custom": {}
+      },
+      {
+          "id": "node_2_1_gy",
+          "layer": 2,
+          "label": "node51",
+          "custom": {}
+      },
+      {
+          "id": "node_2_2_gy",
+          "layer": 2,
+          "label": "node52",
+          "custom": {}
+      },
+      {
+          "id": "node_2_3_gy",
+          "layer": 2,
+          "label": "node53",
+          "custom": {}
+      },
+      {
+          "id": "node_2_4_gy",
+          "layer": 2,
+          "label": "node54",
+          "custom": {}
+      },
+      {
+          "id": "node_2_5_gy",
+          "layer": 2,
+          "label": "node55",
+          "custom": {}
+      },
+      {
+          "id": "node_2_6_gy",
+          "layer": 2,
+          "label": "node56",
+          "custom": {}
+      },
+      {
+          "id": "node_3_0_gy",
+          "layer": 1,
+          "label": "node57",
+          "custom": {}
+      },
+      {
+          "id": "node_3_1_gy",
+          "layer": 1,
+          "label": "node58",
+          "custom": {}
+      },
+      {
+          "id": "node_3_2_gy",
+          "layer": 1,
+          "label": "node59",
+          "custom": {}
+      },
+      {
+          "id": "node_3_3_gy",
+          "layer": 1,
+          "label": "node60",
+          "custom": {}
+      },
+      {
+          "id": "node_3_4_gy",
+          "layer": 1,
+          "label": "node61",
+          "custom": {}
+      },
+      {
+          "id": "node_3_5_gy",
+          "layer": 1,
+          "label": "node62",
+          "custom": {}
+      },
+      {
+          "id": "node_3_6_gy",
+          "layer": 1,
+          "label": "node63",
+          "custom": {}
+      }
+  ],
+  "edges": [
+      {
+          "id": "node_0_0_ss|node_1_0_ss",
+          "source": "node_0_0_ss",
+          "target": "node_1_0_ss",
+          "label": "edge0",
+          "custom": {}
+      },
+      {
+          "id": "node_0_0_ss|node_1_1_ss",
+          "source": "node_0_0_ss",
+          "target": "node_1_1_ss",
+          "label": "edge1",
+          "custom": {}
+      },
+      {
+          "id": "node_0_0_ss|node_1_2_ss",
+          "source": "node_0_0_ss",
+          "target": "node_1_2_ss",
+          "label": "edge2",
+          "custom": {}
+      },
+      {
+          "id": "node_0_0_ss|node_1_3_ss",
+          "source": "node_0_0_ss",
+          "target": "node_1_3_ss",
+          "label": "edge3",
+          "custom": {}
+      },
+      {
+          "id": "node_0_0_ss|node_2_2_ss",
+          "source": "node_0_0_ss",
+          "target": "node_2_2_ss",
+          "label": "edge4",
+          "custom": {}
+      },
+      {
+          "id": "node_0_0_ss|node_1_4_ss",
+          "source": "node_0_0_ss",
+          "target": "node_1_4_ss",
+          "label": "edge5",
+          "custom": {}
+      },
+      {
+          "id": "node_1_0_ss|node_2_0_ss",
+          "source": "node_1_0_ss",
+          "target": "node_2_0_ss",
+          "label": "edge6",
+          "custom": {}
+      },
+      {
+          "id": "node_1_1_ss|node_2_1_ss",
+          "source": "node_1_1_ss",
+          "target": "node_2_1_ss",
+          "label": "edge7",
+          "custom": {}
+      },
+      {
+          "id": "node_1_3_ss|node_2_3_ss",
+          "source": "node_1_3_ss",
+          "target": "node_2_3_ss",
+          "label": "edge8",
+          "custom": {}
+      },
+      {
+          "id": "node_1_4_ss|node_2_4_ss",
+          "source": "node_1_4_ss",
+          "target": "node_2_4_ss",
+          "label": "edge9",
+          "custom": {}
+      },
+      {
+          "id": "node_2_0_ss|node_3_0_ss",
+          "source": "node_2_0_ss",
+          "target": "node_3_0_ss",
+          "label": "edge10",
+          "custom": {}
+      },
+      {
+          "id": "node_2_1_ss|node_3_1_ss",
+          "source": "node_2_1_ss",
+          "target": "node_3_1_ss",
+          "label": "edge11",
+          "custom": {}
+      },
+      {
+          "id": "node_2_2_ss|node_3_2_ss",
+          "source": "node_2_2_ss",
+          "target": "node_3_2_ss",
+          "label": "edge12",
+          "custom": {}
+      },
+      {
+          "id": "node_2_3_ss|node_3_3_ss",
+          "source": "node_2_3_ss",
+          "target": "node_3_3_ss",
+          "label": "edge13",
+          "custom": {}
+      },
+      {
+          "id": "node_2_4_ss|node_3_4_ss",
+          "source": "node_2_4_ss",
+          "target": "node_3_4_ss",
+          "label": "edge14",
+          "custom": {}
+      },
+      {
+          "id": "node_0_0_cjdx|node_1_0_ss",
+          "source": "node_0_0_cjdx",
+          "target": "node_1_0_ss",
+          "label": "edge15",
+          "custom": {}
+      },
+      {
+          "id": "node_0_1_cjdx|node_1_1_ss",
+          "source": "node_0_1_cjdx",
+          "target": "node_1_1_ss",
+          "label": "edge16",
+          "custom": {}
+      },
+      {
+          "id": "node_0_2_cjdx|node_1_2_ss",
+          "source": "node_0_2_cjdx",
+          "target": "node_1_2_ss",
+          "label": "edge17",
+          "custom": {}
+      },
+      {
+          "id": "node_0_3_cjdx|node_1_3_ss",
+          "source": "node_0_3_cjdx",
+          "target": "node_1_3_ss",
+          "label": "edge18",
+          "custom": {}
+      },
+      {
+          "id": "node_0_4_cjdx|node_1_4_ss",
+          "source": "node_0_4_cjdx",
+          "target": "node_1_4_ss",
+          "label": "edge19",
+          "custom": {}
+      },
+      {
+          "id": "node_1_0_cjdx|node_2_0_cjdx",
+          "source": "node_1_0_cjdx",
+          "target": "node_2_0_cjdx",
+          "label": "edge20",
+          "custom": {}
+      },
+      {
+          "id": "node_1_1_cjdx|node_2_1_cjdx",
+          "source": "node_1_1_cjdx",
+          "target": "node_2_1_cjdx",
+          "label": "edge21",
+          "custom": {}
+      },
+      {
+          "id": "node_1_2_cjdx|node_2_2_cjdx",
+          "source": "node_1_2_cjdx",
+          "target": "node_2_2_cjdx",
+          "label": "edge22",
+          "custom": {}
+      },
+      {
+          "id": "node_1_3_cjdx|node_2_3_cjdx",
+          "source": "node_1_3_cjdx",
+          "target": "node_2_3_cjdx",
+          "label": "edge23",
+          "custom": {}
+      },
+      {
+          "id": "node_1_4_cjdx|node_2_4_cjdx",
+          "source": "node_1_4_cjdx",
+          "target": "node_2_4_cjdx",
+          "label": "edge24",
+          "custom": {}
+      },
+      {
+          "id": "node_1_5_cjdx|node_2_5_cjdx",
+          "source": "node_1_5_cjdx",
+          "target": "node_2_5_cjdx",
+          "label": "edge25",
+          "custom": {}
+      },
+      {
+          "id": "node_0_0_gy|node_1_0_gy",
+          "source": "node_0_0_gy",
+          "target": "node_1_0_gy",
+          "label": "edge26",
+          "custom": {}
+      },
+      {
+          "id": "node_0_1_gy|node_1_1_gy",
+          "source": "node_0_1_gy",
+          "target": "node_1_1_gy",
+          "label": "edge27",
+          "custom": {}
+      },
+      {
+          "id": "node_0_2_gy|node_1_2_gy",
+          "source": "node_0_2_gy",
+          "target": "node_1_2_gy",
+          "label": "edge28",
+          "custom": {}
+      },
+      {
+          "id": "node_0_2_gy|node_1_3_gy",
+          "source": "node_0_2_gy",
+          "target": "node_1_3_gy",
+          "label": "edge29",
+          "custom": {}
+      },
+      {
+          "id": "node_1_0_gy|node_2_0_gy",
+          "source": "node_1_0_gy",
+          "target": "node_2_0_gy",
+          "label": "edge30",
+          "custom": {}
+      },
+      {
+          "id": "node_0_0_gy|node_2_1_gy",
+          "source": "node_0_0_gy",
+          "target": "node_2_1_gy",
+          "label": "edge31",
+          "custom": {}
+      },
+      {
+          "id": "node_1_1_gy|node_2_2_gy",
+          "source": "node_1_1_gy",
+          "target": "node_2_2_gy",
+          "label": "edge32",
+          "custom": {}
+      },
+      {
+          "id": "node_0_1_gy|node_2_3_gy",
+          "source": "node_0_1_gy",
+          "target": "node_2_3_gy",
+          "label": "edge33",
+          "custom": {}
+      },
+      {
+          "id": "node_1_2_gy|node_2_4_gy",
+          "source": "node_1_2_gy",
+          "target": "node_2_4_gy",
+          "label": "edge34",
+          "custom": {}
+      },
+      {
+          "id": "node_1_3_gy|node_2_5_gy",
+          "source": "node_1_3_gy",
+          "target": "node_2_5_gy",
+          "label": "edge35",
+          "custom": {}
+      },
+      {
+          "id": "node_0_2_gy|node_2_6_gy",
+          "source": "node_0_2_gy",
+          "target": "node_2_6_gy",
+          "label": "edge36",
+          "custom": {}
+      },
+      {
+          "id": "node_2_0_gy|node_3_0_gy",
+          "source": "node_2_0_gy",
+          "target": "node_3_0_gy",
+          "label": "edge37",
+          "custom": {}
+      },
+      {
+          "id": "node_2_1_gy|node_3_1_gy",
+          "source": "node_2_1_gy",
+          "target": "node_3_1_gy",
+          "label": "edge38",
+          "custom": {}
+      },
+      {
+          "id": "node_2_2_gy|node_3_2_gy",
+          "source": "node_2_2_gy",
+          "target": "node_3_2_gy",
+          "label": "edge39",
+          "custom": {}
+      },
+      {
+          "id": "node_2_3_gy|node_3_3_gy",
+          "source": "node_2_3_gy",
+          "target": "node_3_3_gy",
+          "label": "edge40",
+          "custom": {}
+      },
+      {
+          "id": "node_2_4_gy|node_3_4_gy",
+          "source": "node_2_4_gy",
+          "target": "node_3_4_gy",
+          "label": "edge41",
+          "custom": {}
+      },
+      {
+          "id": "node_2_5_gy|node_3_5_gy",
+          "source": "node_2_5_gy",
+          "target": "node_3_5_gy",
+          "label": "edge42",
+          "custom": {}
+      },
+      {
+          "id": "node_2_6_gy|node_3_6_gy",
+          "source": "node_2_6_gy",
+          "target": "node_3_6_gy",
+          "label": "edge43",
+          "custom": {}
+      }
+  ]
+};
+const colors = ['#F00', '#0f0', '#00f', '#ff0', '#0ff', '#f0f']
+data.nodes.forEach(node => {
+  node.layer *= 2;
+  node.label = `${node.label}|||${node.layer}`
+  node.style = {
+    fill: colors[node.layer]
+  }
+  node.type = 'rect';
+  node.size = [40, 50]
+})
+const div = document.createElement("div");
+div.id = "global-spec";
+document.body.appendChild(div);
+
+describe('#DagreLayout', () => {
+  it('return correct default config', () => {
+    const dagre = new Layouts['dagre']({
+      // nodesepFunc: (d) => {
+      //   return 200;
+      // },
+      nodesep: 1,
+      ranksep: 20,
+      controlPoints: true,
+      fitCenter: true,
+    });
+    dagre.layout(data);
+    const graph = new G6.Graph({
+      container: div,
+      width: 1000,
+      height: 500,
+      modes: {
+        default: ["drag-node", "zoom-canvas", "drag-canvas"],
+      },
+    });
+
+    graph.data(data);
+    graph.render();
+  });
+});