+++ src/layout/types.ts
@@ -7,6 +7,8 @@ export interface Node {
 export interface OutNode extends Node {
   x: number;
   y: number;
+  fx?: number;
+  fy?: number;
   comboId?: string;
   layer?: number; // dagre布局中指定的层级
   _order?: number; // dagre布局中层内排序结果，用于增量布局