+++ src/util/gpu.ts
@@ -52,8 +52,10 @@ export const buildTextureData = (nodes: OutNode[], edges: Edge[]): {
     const e = edges[i];
     const source = getEdgeTerminal(e, 'source');
     const target = getEdgeTerminal(e, 'target');
-    nodeDict[mapIdPos[source]].push(mapIdPos[target]);
-    nodeDict[mapIdPos[target]].push(mapIdPos[source]);
+    if (!isNaN(mapIdPos[source]) && !isNaN(mapIdPos[target])) {
+      nodeDict[mapIdPos[source]].push(mapIdPos[target]);
+      nodeDict[mapIdPos[target]].push(mapIdPos[source]);
+    }
   }
 
   let maxEdgePerVetex = 0;
@@ -61,12 +63,13 @@ export const buildTextureData = (nodes: OutNode[], edges: Edge[]): {
     const offset: number = dataArray.length;
     const dests = nodeDict[i];
     const len = dests.length;
+    console.log('dests', dests, len)
     dataArray[i * 4 + 2] = offset;
-    dataArray[i * 4 + 3] = dests.length;
-    maxEdgePerVetex = Math.max(maxEdgePerVetex, dests.length);
+    dataArray[i * 4 + 3] = len;
+    maxEdgePerVetex = Math.max(maxEdgePerVetex, len);
     for (let j = 0; j < len; ++j) {
-    const dest = dests[j];
-    dataArray.push(+dest);
+      const dest = dests[j];
+      dataArray.push(+dest);
     }
   }